+++ TractorMod/release-notes.md
@@ -1,5 +1,6 @@
 ﻿# Release notes
 ## Upcoming release
+* Fixed heavy tappers not detected consistently due to a game bug.
 * Fixed tractor watering can not playing any sound.
 * Fixed crash if mod is installed incorrectly and its textures files don't exist.